# openapi.yaml
openapi: 3.1.0
info:
  title: Aranis API
  version: "1.0.0"
  summary: Third-party risk, cyber risk and privacy data from your Aranis workspace.
  description: |
    The Aranis API gives an Enterprise workspace programmatic read access to its
    suppliers, assessments, risks, action plans and audit trail, plus a small set of
    writes for keeping an external inventory in sync.

    It is a **server-to-server** API. There is no CORS header on any response and no
    preflight is answered, so it cannot be called from a browser — a key belongs on
    your backend, never in a page.

    ## Authentication

    Send your key as a bearer token:

    ```
    Authorization: Bearer ara_live_...
    ```

    Keys are created in the Aranis app under **Settings → Integrations**. The secret is
    shown once, at creation. Each key carries an explicit set of scopes and belongs to
    exactly one workspace — the workspace is always derived from the key, which is why
    no endpoint accepts an organization identifier as input.

    ## Availability

    The API is part of the Enterprise plan. A valid key on a workspace that is not on
    Enterprise receives `403 plan_upgrade_required` rather than `401`: the credential is
    fine, the entitlement is not.

    ## Personal data

    `pii:read` is additive. It grants nothing on its own — it un-redacts personal fields
    on resources some other scope already allows. Without it, those fields are **absent
...

...

...

  contact:
    name: Aranis Support
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /assessments/{id}:
    get:
      tags: [Assessments]
      operationId: getAssessment
      summary: Retrieve an assessment
      security:
        - bearerAuth: [assessments:read]
      parameters:
        - $ref: '#/components/parameters/resourceId'
      responses:
        '200':
          description: The assessment.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Assessment' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /assessments/{id}/gaps:
    get:
      tags: [Assessments]
      operationId: listAssessmentGaps
      summary: List an assessment's gaps
      description: |
        The controls this assessment did not fully meet, as a derived list.

        Each gap carries the code and title of its own control. There is no way to page,
        filter or enumerate the control catalogue through this endpoint, and nothing here
        reveals a control the assessment did not fail. The supplier's raw answer text is
        never returned.

        `evidence_status` is always present. The `evidence` array — file name, MIME type,
        SHA-256, review status, score — appears only when the key also holds
        `evidence:read`. Evidence file contents never leave the platform.
      security:
        - bearerAuth: [assessments:read]
      parameters:
        - $ref: '#/components/parameters/resourceId'
        - name: severity
          in: query
          schema: { type: string, enum: [low, medium, high] }
      responses:
        '200':
          description: |
            Every gap in the assessment, ordered by control code. A single assessment
            fits in one response, so this collection is never paged.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Collection'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Gap' }
              example:
                data:
                  - control_code: ARC08
                    control_title: Segregation of duties in production access
                    gap_description: Without segregation, a single operator can both deploy and approve changes.
                    domain: cyber
                    status: partially_met
                    severity: high
                    evidence_status: missing
                    evidence: []
                meta: { has_more: false, next_cursor: null, feature_status: available }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }

  # ── Reports ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <key>`. Keys are created in the Aranis app under
        Settings → Integrations and carry an explicit scope set. The workspace is always
        derived from the key.
  parameters:
...
    resourceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: The request was malformed — an unknown filter value, a bad cursor, an out-of-range limit.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: |
        Missing or invalid credentials. Every failure mode — unknown, revoked, expired —
        returns this same response; distinguishing them would tell an attacker which keys
        exist.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: invalid_api_key
              message: The API key provided is not valid.
              request_id: req_01J8Z9K2M4N6P8Q0R2S4T6V8
    Forbidden:
      description: |
        The credential is valid but not entitled. Either the workspace is not on
        Enterprise (`plan_upgrade_required`) or the key lacks the scope this endpoint
        requires (`insufficient_scope`).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            planUpgradeRequired:
              value:
                error:
                  code: plan_upgrade_required
                  message: The Aranis API is available on the Enterprise plan. Contact sales to enable it.
                  request_id: req_01J8Z9K2M4N6P8Q0R2S4T6V8
            insufficientScope:
              value:
                error:
                  code: insufficient_scope
                  message: This API key is missing the 'context:read' scope required for this endpoint.
                  request_id: req_01J8Z9K2M4N6P8Q0R2S4T6V8
                  details: { required_scope: context:read }
    NotFound:
      description: |
        No such resource in this workspace. A valid id belonging to another workspace also
        returns 404 — the API never confirms that a resource exists elsewhere.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    RateLimited:
      description: |
        Too many requests. `Retry-After` gives the seconds to wait.

        Limits are per workspace: 60 req/s burst, 600 req/min sustained, 5,000 req/hour.
        `GET /reports/{id}/pdf` additionally carries its own 60/hour ceiling.
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds to wait before retrying.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              enum:
                - invalid_request
                - invalid_api_key
                - plan_upgrade_required
                - insufficient_scope
                - resource_not_found
                - conflict
                - validation_failed
                - rate_limit_exceeded
                - internal_error
            message:
              type: string
              description: A human-readable explanation. Safe to log; do not parse.
            request_id:
              type: string
              description: |
                Also returned as the `X-Request-Id` header on every response, including
                successful ones. This is what support will ask for.
            details:
              type: object
              additionalProperties: true
    CollectionMeta:
      type: object
      required: [has_more, next_cursor, feature_status]
      properties:
        has_more:
          type: boolean
          description: Whether another page exists.
        next_cursor:
          type: [string, 'null']
          description: Pass as `cursor` to fetch the next page. Null on the last page.
        feature_status:
          type: string
          enum: [available, not_configured, coming_soon]
          description: |
            Why a collection is empty, when it is.

            - `available` — the feature is active. An empty `data` means you have no rows.
            - `not_configured` — the feature exists but this workspace has not set it up.
            - `coming_soon` — not released yet.

            An empty collection is never a 404 and never a 500.
        message:
          type: string
          description: Present when `feature_status` is not `available`.
    Collection:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: {}
        meta: { $ref: '#/components/schemas/CollectionMeta' }

    # ── enums ──────────────────────────────────────────────────────────────────
...
    VendorProfile:
      type: string
      enum: [P0, P1, P2, P3]
      description: Assessment depth. P0 is a public-evidence review; P3 is the full control pool.
...
    AssessmentStatus:
      type: string
      enum: [draft, sent, in_progress, paused, completed, expired]
    AssessmentType:
      type: string
      enum: [self, supplier]
    AssessmentDecision:
      type: string
      enum: [approved, approved_with_caveats, rejected]
...
    Assessment:
      type: object
      properties:
        id: { type: string, format: uuid }
        supplier_id:
          type: [string, 'null']
          format: uuid
          description: "Null for self-assessments."
        assessment_type: { $ref: '#/components/schemas/AssessmentType' }
        status: { $ref: '#/components/schemas/AssessmentStatus' }
        vendor_profile: { oneOf: [{ $ref: '#/components/schemas/VendorProfile' }, { type: 'null' }] }
        deadline: { type: [string, 'null'], format: date-time }
        completed_at: { type: [string, 'null'], format: date-time }
        validity_days: { type: [integer, 'null'] }
        expires_at: { type: [string, 'null'], format: date-time }
        scores:
          type: [object, 'null']
          additionalProperties: true
          description: |
            Computed scores. `overall` is the headline; per-domain keys (`cyber`,
            `privacy`, `bcm`, `ai`) and `business_risk` accompany it. Null until the
            assessment completes.
        result_note: { type: [string, 'null'] }
        scanner_used: { type: [boolean, 'null'] }
        decision: { oneOf: [{ $ref: '#/components/schemas/AssessmentDecision' }, { type: 'null' }] }
        decision_reason_text: { type: [string, 'null'] }
        decided_at: { type: [string, 'null'], format: date-time }
        is_shared: { type: boolean }
        sharing_enabled: { type: [boolean, 'null'] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Gap:
      type: object
      description: A control the assessment did not fully meet.
      properties:
        control_code:
          type: string
          description: "The control's public code, e.g. `ARC08`."
        control_title: { type: [string, 'null'] }
        gap_description:
          type: [string, 'null']
          description: "What this gap means in practice."
        domain: { type: [string, 'null'] }
        status: { type: string, enum: [not_met, partially_met] }
        severity:
          type: string
          enum: [low, medium, high]
          description: "Derived from the control's canonical CVSS."
        evidence_status:
          type: string
          enum: [provided, pending_review, missing, unknown]
          description: '`unknown` means the key lacks `evidence:read`, so evidence was not consulted.'
        evidence:
          type: array
          description: 'Metadata only, and only with `evidence:read`. File contents never leave the platform.'
          items:
            type: object
            properties:
              file_name: { type: [string, 'null'] }
              mime_type: { type: [string, 'null'] }
              sha256: { type: [string, 'null'] }
              status: { type: [string, 'null'] }
              score: { type: [number, 'null'] }
              uploaded_at: { type: [string, 'null'], format: date-time }
...